ecore-evas: Add ecore_evas function used to set cloned outputs
This patch set adds a new API function for ecore_evas_output_clone_set
which can be used from Enlightenment in order to set one output as a
clone of another. This is used in the wl_drm module when Screen
Settings is configured to clone or unclone an output.