efl_ui_box: early return on errors
when efl_ui_widget_sub_object_add does fail, then we should definitly
return, and not try to add it to the box. This preserves a giant amount
of errors. Additionally, it applies the same rules we have in the
internals of the widgets to the widget, like not applying our self to
our own container.