efl_ui_box: error out when a widget is already part of the container
code normally does not check if the packing of a widget was successfull not.
This means, that a case where something was accidently added twice to
the container is not easily discovered. With printing here an error we
enable error checking testcases the ability to check if the assembly of
a partial UI was successfull or not.