tests: align timeout value across test suites
We had a big variety of timeout values spread across our test suites.
Ranging from 30s (if not defined, meson default) to 600s.

While the meson default with 30s might be a bit tight depending on the
system executing the test suite ( I have seen elementary and evas going
up to 28s here on my laptop) 10 minutes is surely to long and we would
like to know if a test suite takes that long.

The 60s I propose here are simply taking the longest times I have seen
here so fare and double it to allow for slow or heavily loaded systems.

Our total test time is around 1-2 minutes right now and we surely want
to keep it below 2, better aiming for below 1 minute. Having an early
indicator on time heavy tests with a sane timeout value helps us to get
them identifies and hopefully fixed quickly.