csharp WIP: Make events work with dissociated wrappers
This is an alternative approach to the fix we are working on for T7800

This commit keeps the C# wrappers dissociated from the Eo instances
(i.e. the wrappers may die while the Eo is still alive).

The events work by creating new wrappers upon event callback and keeping
the dictionary with the event callback alive for the duraction of the Eo
lifetime.

Still to do:

- Overload '==' ?
- Dispose the dictionary when the object emits "DEL"
- Attach the existing dictionary when creating the wrapper from the
  constructor receiving the raw Pointer.