Rename interface Efl.Ui.Direction to Efl.Ui.Directed
Summary:
Interfaces should be Adjectives: A layout has a direction, and therefore it is
Directed. It is NOT a direction.
c_prefix is used so the method names do not change in C land.

Related to T7863

Test Plan:
Everything still builds and passes tests, both with meson and autotools.
No functional changes.

Reviewers: zmike, bu5hm4n, cedric

Subscribers: #reviewers, #committers

Tags: #efl

Differential Revision: https://phab.enlightenment.org/D8898